feat(transfer): refactor code to prepare adding transfer flow
- restructure files to allow more flexibility
- add bare-bones implementation for ml-thirdparty-client
- add bare-bones implementation for ml-thirdparty-simulator
- create internal plugins for mojaloop client and simulator
