+++ package.json
@@ -60,8 +60,9 @@
   "devDependencies": {
     "@commitlint/cli": "^9.0.1",
     "@commitlint/config-conventional": "^9.0.1",
-    "@mojaloop/api-snippets": "git+https://github.com/stevenwjy/api-snippets.git#1e3a7bdcb2f0468d57d87c2c922831a931d183c0",
+    "@mojaloop/api-snippets": "git+https://github.com/stevenwjy/api-snippets.git#446972b2d7c8aaa6cc000ec4eb738e9b9ec14eb9",
     "@types/convict": "^5.2.1",
+    "@types/faker": "^4.1.12",
     "@types/hapi": "^18.0.3",
     "@types/hapi__hapi": "^19.0.3",
     "@types/hapi__inert": "^5.2.0",
@@ -102,6 +103,8 @@
     "@hapi/inert": "^6.0.1",
     "@hapi/vision": "^6.0.0",
     "convict": "^6.0.0",
+    "faker": "^4.1.0",
+    "firebase-admin": "^8.13.0",
     "openapi-backend": "^3.5.1",
     "winston": "^3.3.3"
   }
