+++ src/server/handlers/openapi/app/consents/{ID}/generateChallenge.ts
