+++ src/server/handlers/openapi/app/parties/{Type}/{ID}.ts
