+++ src/server/handlers/openapi/mojaloop/index.ts
@@ -23,11 +23,29 @@
  --------------
  ******/
 
-import { Request, ResponseToolkit } from '@hapi/hapi'
-import { Handler, Context } from 'openapi-backend'
-import { logger } from '~/shared/logger'
+import * as Health from '../health'
 
-export const post: Handler = async (context: Context, request: Request, h: ResponseToolkit) => {
-  logger.logRequest(context, request, h)
-  return h.response().code(202)
+import * as MojaloopAuthorizations from './authorizations'
+import * as MojaloopConsents from './consents'
+import * as MojaloopConsentsById from './consents/{ID}'
+import * as MojaloopConsentRequestsById from './consentRequests/{ID}'
+import * as MojaloopParticipants from './participants'
+import * as MojaloopParticipantsError from './participants/error'
+import * as MojaloopPartiesByTypeAndId from './parties/{Type}/{ID}'
+import * as MojaloopPartiesByTypeAndIdError from './parties/{Type}/{ID}/error'
+import * as MojaloopTransfersById from './transfers/{ID}'
+
+export const apiHandlers = {
+  getHealth: Health.get,
+
+  postAuthorizations: MojaloopAuthorizations.post,
+  postConsents: MojaloopConsents.post,
+  putConsentsById: MojaloopConsentsById.put,
+  deleteConsentsById: MojaloopConsentsById.remove,
+  putConsentRequestsById: MojaloopConsentRequestsById.put,
+  putParticipants: MojaloopParticipants.put,
+  putParticipantsError: MojaloopParticipantsError.put,
+  putPartiesByTypeAndId: MojaloopPartiesByTypeAndId.put,
+  putPartiesByTypeAndIdError: MojaloopPartiesByTypeAndIdError.put,
+  putTransfersById: MojaloopTransfersById.put,
 }
