+++ src/server/handlers/openapi/mojaloop/participants/error.ts
