+++ src/server/handlers/openapi/mojaloop/parties/{Type}/{ID}/error.ts
