+++ src/server/plugins/internal/mojaloopSimulator.ts
@@ -2,7 +2,7 @@
  License
  --------------
  Copyright © 2020 Mojaloop Foundation
- The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the 'License') and you may not use these files except in compliance with the License. You may obtain a copy of the License at
+ The Mojaloop files are made available by the Mojaloop Foundation under the Apache License, Version 2.0 (the 'License') and you may not use these files except in compliance with the License. You may obtain a copy of the License at
  http://www.apache.org/licenses/LICENSE-2.0
  Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  Contributors
@@ -23,11 +23,20 @@
  --------------
  ******/
 
-import { Request, ResponseToolkit } from '@hapi/hapi'
-import { Handler, Context } from 'openapi-backend'
-import { logger } from '~/shared/logger'
+import { Plugin, Server } from '@hapi/hapi'
+import { Simulator } from '~/shared/ml-thirdparty-simulator'
 
-export const put: Handler = async (context: Context, request: Request, h: ResponseToolkit) => {
-  logger.logRequest(context, request, h)
-  return h.response().code(200)
+export interface MojaloopSimulatorOpts {
+  vhost?: string
+}
+
+export const MojaloopSimulator: Plugin<MojaloopSimulatorOpts> = {
+  name: 'MojaloopSimulator',
+  version: '1.0.0',
+  register: (server: Server, opts: MojaloopSimulatorOpts) => {
+    server.app.mojaloopClient.simulator = new Simulator(
+      server,
+      { ...opts },
+    )
+  }
 }
